pyproject.toml
[project]
name = "hardware"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.3.0",
]
